pyproject.toml
[project]
name = "finquery"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "chromadb>=1.3.5",
    "openai>=2.8.1",
    "pandas>=2.3.3",
    "pdfplumber>=0.11.8",
    "python-dotenv>=1.2.1",
    "tabulate>=0.9.0",
    "tiktoken>=0.12.0",
]
